requirements.txt
torch
torchvision
transformers
accelerate
decord; platform_machine != 'arm64'
eva-decord; platform_machine == 'arm64'
gradio
huggingface_hub
imageio
importlib_metadata
matplotlib
moviepy==1.0.3
numba
numpy
noise
omegaconf
openai
opencv-python>=4.4.0.46
openpyxl
pandas
pillow
portalocker
protobuf
python-dotenv
PyAV
qwen-vl-utils[decord]==0.0.8
requests
rich
scipy
sentencepiece
setuptools
scikit-image
sty
tabulate
tiktoken
timeout-decorator
tqdm
typing_extensions
ultralytics
validators
xlsxwriter
